Move prebid server variants to lib file
